Show sick day follow up question experimentation (#33)
* Add FiveOptionsView

* Revert "Incorporating new Sick Day protocol"

This reverts commit cfbcd7c508074ce4d2cf3698d9fdd987e7703994.

* Configure FourOptions and FiveOptions views

Added initial question IDs for four and five-option question types, together with the views

* Refactor FourOptionsAnswer and update selection handling

Reworked the FourOptionsAnswer enum to use distinct cases for DKA, high blood sugar, low blood sugar, and not sure, each carrying a ChildIssue value. Updated FourOptionsView and GetHelpViewController to handle these new cases and improved selection logic. Also fixed an issue in HomeViewController where the wrong question was being passed to GetHelpViewController.

* Show follow up question experimentation

Currently, we are able to show the follow up question, in this case for testing purposes, I chose the

* proposal for passing result from view to parent;

* Add follow up question view in parent question type

Able to have a follow up question in the TwoOptionsView. In this cases, started by adding another TwoOptionsFollowUpView

---------

Co-authored-by: yago1994 <yagoaral@hotmail.com>